fix(tests): organize folders structure
